+++ packages/benches/basic/secondaryIndexes.bench.ts
@@ -1,7 +1,7 @@
 import { bench } from '@arktype/attest';
 import { a, ClientSchema } from '@aws-amplify/data-schema';
 
-bench('secondary index', () => {
+bench('secondary index of model has only ModelFields', () => {
   const s = a.schema({
     Post: a
       .model({
@@ -17,4 +17,86 @@ bench('secondary index', () => {
   });
 
   type _ = ClientSchema<typeof s>;
-}).types([51984, 'instantiations']);
+}).types([52450, 'instantiations']);
+
+bench(
+  'secondary index without using enum field of a model that has enum field',
+  () => {
+    const s = a.schema({
+      Post: a
+        .model({
+          title: a.string().required(),
+          description: a.string(),
+          status: a.enum(['published', 'reviewing']),
+        })
+        .secondaryIndexes((index) => [
+          index('title').sortKeys(['description']).queryField('myFavIdx'),
+          index('description'),
+        ]),
+    });
+
+    type _ = ClientSchema<typeof s>;
+  },
+).types([56104, 'instantiations']);
+
+bench(
+  'secondary index with using enum field of a model that has enum field',
+  () => {
+    const s = a.schema({
+      Post: a
+        .model({
+          title: a.string().required(),
+          description: a.string(),
+          status: a.enum(['published', 'reviewing']),
+        })
+        .secondaryIndexes((index) => [
+          index('status').sortKeys(['description']).queryField('myFavIdx'),
+          index('description'),
+        ]),
+    });
+
+    type _ = ClientSchema<typeof s>;
+  },
+).types([56129, 'instantiations']);
+
+bench(
+  'secondary index without using enum field of a model that has enum ref field',
+  () => {
+    const s = a.schema({
+      Status: a.enum(['published', 'reviewing']),
+      Post: a
+        .model({
+          title: a.string().required(),
+          description: a.string(),
+          status: a.ref('Status'),
+        })
+        .secondaryIndexes((index) => [
+          index('title').sortKeys(['description']).queryField('myFavIdx'),
+          index('description'),
+        ]),
+    });
+
+    type _ = ClientSchema<typeof s>;
+  },
+).types([54178, 'instantiations']);
+
+bench(
+  'secondary index with using enum field of a model that has enum ref field',
+  () => {
+    const s = a.schema({
+      Status: a.enum(['published', 'reviewing']),
+      Post: a
+        .model({
+          title: a.string().required(),
+          description: a.string(),
+          status: a.ref('Status'),
+        })
+        .secondaryIndexes((index) => [
+          index('status').sortKeys(['description']).queryField('myFavIdx'),
+          index('description'),
+        ]),
+    });
+
+    type _ = ClientSchema<typeof s>;
+  },
+).types([54241, 'instantiations']);