+++ packages/benches/p50/p50-prod.bench.ts
@@ -486,7 +486,7 @@ bench('prod p50', () => {
         ipAddress: a.ipAddress(),
       })
       .authorization((allow) => [
-        allow.authenticated('iam').to(['read']),
+        allow.authenticated('identityPool').to(['read']),
         allow.owner(),
       ]),
     // Model #23:
@@ -521,7 +521,7 @@ bench('prod p50', () => {
         ipAddress: a.ipAddress(),
       })
       .authorization((allow) => [
-        allow.authenticated('iam').to(['read']),
+        allow.authenticated('identityPool').to(['read']),
         allow.owner(),
       ]),
     // Model #24:
@@ -556,7 +556,7 @@ bench('prod p50', () => {
         ipAddress: a.ipAddress(),
       })
       .authorization((allow) => [
-        allow.authenticated('iam').to(['read']),
+        allow.authenticated('identityPool').to(['read']),
         allow.owner(),
       ]),
     // Model #25:
@@ -591,7 +591,7 @@ bench('prod p50', () => {
         ipAddress: a.ipAddress(),
       })
       .authorization((allow) => [
-        allow.authenticated('iam').to(['read']),
+        allow.authenticated('identityPool').to(['read']),
         allow.owner(),
       ]),
     // [Global authorization rule]
@@ -1075,7 +1075,7 @@ bench('prod p50 w/ client types', () => {
           ipAddress: a.ipAddress(),
         })
         .authorization((allow) => [
-          allow.authenticated('iam').to(['read']),
+          allow.authenticated('identityPool').to(['read']),
           allow.owner(),
         ]),
       // Model #23:
@@ -1110,7 +1110,7 @@ bench('prod p50 w/ client types', () => {
           ipAddress: a.ipAddress(),
         })
         .authorization((allow) => [
-          allow.authenticated('iam').to(['read']),
+          allow.authenticated('identityPool').to(['read']),
           allow.owner(),
         ]),
       // Model #24:
@@ -1145,7 +1145,7 @@ bench('prod p50 w/ client types', () => {
           ipAddress: a.ipAddress(),
         })
         .authorization((allow) => [
-          allow.authenticated('iam').to(['read']),
+          allow.authenticated('identityPool').to(['read']),
           allow.owner(),
         ]),
       // Model #25:
@@ -1180,7 +1180,7 @@ bench('prod p50 w/ client types', () => {
           ipAddress: a.ipAddress(),
         })
         .authorization((allow) => [
-          allow.authenticated('iam').to(['read']),
+          allow.authenticated('identityPool').to(['read']),
           allow.owner(),
         ]),
       // [Global authorization rule]
@@ -1667,7 +1667,7 @@ bench('prod p50 combined w/ client types', () => {
           ipAddress: a.ipAddress(),
         })
         .authorization((allow) => [
-          allow.authenticated('iam').to(['read']),
+          allow.authenticated('identityPool').to(['read']),
           allow.owner(),
         ]),
       // Model #23:
@@ -1702,7 +1702,7 @@ bench('prod p50 combined w/ client types', () => {
           ipAddress: a.ipAddress(),
         })
         .authorization((allow) => [
-          allow.authenticated('iam').to(['read']),
+          allow.authenticated('identityPool').to(['read']),
           allow.owner(),
         ]),
       // Model #24:
@@ -1737,7 +1737,7 @@ bench('prod p50 combined w/ client types', () => {
           ipAddress: a.ipAddress(),
         })
         .authorization((allow) => [
-          allow.authenticated('iam').to(['read']),
+          allow.authenticated('identityPool').to(['read']),
           allow.owner(),
         ]),
       // Model #25:
@@ -1772,7 +1772,7 @@ bench('prod p50 combined w/ client types', () => {
           ipAddress: a.ipAddress(),
         })
         .authorization((allow) => [
-          allow.authenticated('iam').to(['read']),
+          allow.authenticated('identityPool').to(['read']),
           allow.owner(),
         ]),
       // [Global authorization rule]