+++ packages/data-schema/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aws-amplify/data-schema
 
+## 0.18.3
+
+### Patch Changes
+
+- d473fbb: feat: making enum field or ref of enum field eligible to be PK or SK of 2nd-ary index
+
+## 0.18.2
+
+### Patch Changes
+
+- b3fffcf: feat: rename iam -> identityPool
+- Updated dependencies [b3fffcf]
+  - @aws-amplify/data-schema-types@0.10.1
+
 ## 0.18.1
 
 ### Patch Changes